+++ proxy/middleware/instrument.go
@@ -23,6 +23,12 @@ func Instrument(h http.Handler) http.Handler {
 			start := time.Now()
 			defer func() {
 				code := strconv.Itoa(lrw.Status())
+				/*
+				We can't statically define these patterns
+				because we don't know all the combinations
+				we'll see. So we dynamically create/increment
+				paths
+				*/
 				reqStr := fmt.Sprintf("http_requests_total{method=%q,path=%q,status_code=%q}", method, path, code)
 				vmmetrics.GetOrCreateCounter(reqStr).Inc()
 				takenStr := fmt.Sprintf("http_request_time_secs_total{method=%q,path=%q,status_code=%q}", method, path, code)